мелкие правки и перенос констант
