+++ src/components/section.js
@@ -1,7 +1,4 @@
 export default class Section {
-    //items - массив данных, которые добавляем на страницу, сами данные получим от Api
-    //renderer - функция отвечающая за отрисовку данных
-    // selectorContainer - контейнер куда добавляются все отрисованные карточки
     constructor({items, renderer}, selectorContainer)  {
         this._arrayCards = items;
         this._renderer = renderer;
